requirements.txt
sqlalchemy
psycopg2
pandas
dotenv
pyyaml
scikit-learn==1.2.2
numpy==1.24.4